requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pillow==10.4.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
python-multipart==0.0.9
requests==2.32.3
pydantic==2.9.2
motor==3.5.1
pymongo==4.5.0
bcrypt==4.1.3
PyJWT==2.9.0
pydantic-settings==2.4.0
email-validator==2.1.0
